Move logic from Booking to Space Booking for reuse
This commit extracts the logic used to extract ‘Management Info’ from delius booking imports into a common service such that it can be reused for a new seed jobs that creates `SpaceBooking` for delius referrals that don’t have an existing booking in CAS1.
